# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PrintGrid API
  version: 0.2.0-contract-first
  description: >
    HỢP ĐỒNG API KHÓA NGÀY 27/09 (nhiệm vụ "Thống nhất hợp đồng API và nền móng repo chung").
    Quy tắc: đổi schema/endpoint TRƯỚC hết phải sửa file này, PR qua chủ vỏ (ghế Nền tảng),
    rồi mới code. Cột x-milestone chỉ tuần dự kiến có mã thật:
    existing = đã chạy, T1..T8 = theo kế hoạch 1 trang.
servers:
  - url: http://localhost:5000/api/v1
    description: API trực tiếp từ docker-compose (host 5000 → container 8080) — smoke test 27/09
  - url: http://localhost:3000/api/v1
    description: qua nginx frontend (proxy /api — smoke test 27/09)
  - url: https://staging.printgrid.dev/api/v1
    description: staging (từ T7)
...
security:
  - bearerAuth: []
paths:
...
  /me/addresses/{addressId}:
    put:
      tags: [customer]
      summary: Cập nhật địa chỉ và chọn mặc định
      parameters: [ { name: addressId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AddressInput' } } }
      responses: { '200': { description: AddressDto }, '404': { description: Không tìm thấy } }
    delete:
      tags: [customer]
      x-milestone: T1
      summary: Xóa địa chỉ
      parameters: [ { name: addressId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '204': { description: Đã xóa }
        '409': { description: Còn đơn chưa giao đang dùng địa chỉ này }
...
  /labs/{labId}/inventory:
    parameters: [ { name: labId, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:  { tags: [lab], summary: Tồn kho theo vật liệu và màu, responses: { '200': { description: "MaterialStockDto[]" } } }
    put:  { tags: [lab], summary: Kiểm kê/cấu hình ngưỡng tồn kho, responses: { '200': { description: "MaterialStockDto" } } }
  /labs/{labId}/inventory/adjustments:
    parameters: [ { name: labId, in: path, required: true, schema: { type: string, format: uuid } } ]
    post: { tags: [lab], summary: Nhập hoặc xuất kho có mã giao dịch và số dư liên tục, responses: { '200': { description: "MaterialStockDto" } } }
  /labs/{labId}/inventory/transactions:
    parameters: [ { name: labId, in: path, required: true, schema: { type: string, format: uuid } } ]
    get: { tags: [lab], summary: Lịch sử giao dịch kho, responses: { '200': { description: "StockTransactionDto[]" } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  schemas:
...
    AddressInput:
      type: object
      required: [label, recipientName, phoneNumber, street, city, isDefault]
      properties:
        label: { type: string, maxLength: 50 }
        recipientName: { type: string, maxLength: 200 }
        phoneNumber: { type: string, maxLength: 32 }
        street: { type: string, maxLength: 300 }
        ward: { type: string, maxLength: 100 }
        district: { type: string, maxLength: 100 }
        city: { type: string, maxLength: 100 }
        postalCode: { type: string, maxLength: 20 }
        country: { type: string, minLength: 2, maxLength: 2, default: VN }
        isDefault: { type: boolean }
...